Enable Gradle Configuration Cache on GitHub Actions
In #899, the `cache-encryption-key` argument was removed from the `gradle/actions/setup-gradle` action to avoid using secrets.
However, this secret is optional: if the value is not provided, the cache won't be read or updated.
This means that PRs created from the main repository will take advantage of Gradle's Configuration Cache, while PRs from forks won't.